// src/components/llm-components/types.ts
export type ButtonVariant = "primary" | "secondary" | "outline" | "ghost";
export type ButtonSize = "sm" | "md" | "lg";
// ...
export interface ButtonComponent {
  type: "button";
  label?: string;
  action?: string;
  variant?: ButtonVariant;
  size?: ButtonSize;
}